refactor: consolidate type definitions and improve TypeScript config
- Move User type from API to shared types package for better reusability
- Consolidate Express type definitions into single declaration file
- Update TypeScript configuration to use standard type resolution
- Apply consistent code formatting across type files
