+++ apps/api/src/middleware/auth.ts
@@ -1,3 +1,4 @@
+/// <reference path="../index.d.ts" />
 import type { NextFunction, Request, Response } from 'express';
 import jwt from 'jsonwebtoken';
