+++ packages/types/src/UserProducts.ts
@@ -1,6 +1,6 @@
-import { CashAccount } from './CashAccount';
-import { Deposit } from './Deposit';
-import { IndexedFund } from './IndexedFund';
+import type { CashAccount } from "./CashAccount";
+import type { Deposit } from "./Deposit";
+import type { IndexedFund } from "./IndexedFund";
 
 export interface UserProducts {
   cashAccounts: CashAccount[];
