+++ packages/types/src/index.ts
@@ -1,5 +1,5 @@
-export { Product } from './Product';
-export { CashAccount } from './CashAccount';
-export { Deposit } from './Deposit';
-export { IndexedFund } from './IndexedFund';
-export { UserProducts } from './UserProducts';
+export { Product } from "./Product";
+export { CashAccount } from "./CashAccount";
+export { Deposit } from "./Deposit";
+export { IndexedFund } from "./IndexedFund";
+export { UserProducts } from "./UserProducts";
